+++ example/lib/main.dart
@@ -8,7 +8,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -24,13 +24,13 @@ class MyApp extends StatelessWidget {
         /* dark theme settings */
       ),
       debugShowCheckedModeBanner: false,
-      home: const MyHomePage(title: 'Dropdwon Time picker Demo'),
+      home: const MyHomePage(title: 'Dropdown Time Picker Demo'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
+  const MyHomePage({super.key, required this.title});
 
   final String title;